Add files via upload
I design mini-project using Next js and MySQL consisting of two pages: one page to 
input and store the data of schools, the other page to fetch and display the data of the schools.
page1: addSchool : I create a form in next js (using react-hook-form) using which a user can 
enter and store the data of a school in the `schools` table. 
page2: showSchool: I create a page that shows the list of schools in the database in the way an 
ecommerce website shows products. Only the School’s name, address, city and image needs to be 
visible here.